All these new flags only apply if the --headless flag is specified. Fixes [https://trac.osgeo.org/mapguide/ticket/2768]
........
Merged revision(s) 8967, 8972-8973 from branches/3.1/vagrant:
Fix release label not being properly stamped onto ubuntu install script as it was looking for a default label that doesn't exist.
........
Update ubuntu install script:
 - Add extra pre-requisite packages that were discovered to be required when testing on a clean barebones Ubuntu server installation.
........
Remove libc6 requirement from mginstallubuntu.sh. This package will always be present.
........

........


git-svn-id: http://svn.osgeo.org/mapguide/branches/3.0@9170 a442ad82-b40c-0410-ae44-e6d7bd6e4a36

+++ vagrant/clean.bat
@@ -6,50 +6,70 @@ if exist updated rd /S /Q updated
 if exist filelist rd /S /Q filelist
 if exist fdosdk_filename del /F fdosdk_filename
 if exist fdosdk_rev del /F fdosdk_rev
-del /F fdosdk*.tar.gz
+if exist fdosdk*.tar.gz del /F fdosdk*.tar.gz
 popd
-pushd centos\src_update
+pushd centos6\src_update
 call vagrant destroy -f
 popd
-pushd ubuntu\x86
-SET COMPONENT=Ubuntu Build 32-bit
-echo [clean]: MapGuide Ubuntu build 32-bit
+pushd ubuntu12\x86
+SET COMPONENT=Ubuntu 12 Build 32-bit
+echo [clean]: %COMPONENT%
+call vagrant destroy -f
 if exist build rd /S /Q build
 if exist filelist rd /S /Q filelist
 if exist *.sh del *.sh
 if exist *.pl del *.pl
 if exist *.txt del *.txt
-call vagrant destroy -f
 popd
-pushd centos\x86
-SET COMPONENT=CentOS Build 32-bit
-echo [clean]: MapGuide CentOS build 32-bit
+pushd ubuntu14\x86
+SET COMPONENT=Ubuntu 14 Build 32-bit
+echo [clean]: %COMPONENT%
+call vagrant destroy -f
 if exist build rd /S /Q build
 if exist filelist rd /S /Q filelist
 if exist *.sh del *.sh
 if exist *.pl del *.pl
 if exist *.txt del *.txt
-call vagrant destroy -f
 popd
-pushd ubuntu\x64
-SET COMPONENT=Ubuntu Build 64-bit
-echo [clean]: MapGuide Ubuntu build 64-bit
+pushd centos6\x86
+SET COMPONENT=CentOS 6 Build 32-bit
+echo [clean]: %COMPONENT%
+call vagrant destroy -f
 if exist build rd /S /Q build
 if exist filelist rd /S /Q filelist
 if exist *.sh del *.sh
 if exist *.pl del *.pl
 if exist *.txt del *.txt
-call vagrant destroy -f
 popd
-pushd centos\x64
-SET COMPONENT=CentOS Build 64-bit
-echo [clean]: MapGuide CentOS build 64-bit
+pushd ubuntu12\x64
+SET COMPONENT=Ubuntu 12 Build 64-bit
+echo [clean]: %COMPONENT%
+call vagrant destroy -f
 if exist build rd /S /Q build
 if exist filelist rd /S /Q filelist
 if exist *.sh del *.sh
 if exist *.pl del *.pl
 if exist *.txt del *.txt
+popd
+pushd ubuntu14\x64
+SET COMPONENT=Ubuntu 14 Build 64-bit
+echo [clean]: %COMPONENT%
 call vagrant destroy -f
+if exist build rd /S /Q build
+if exist filelist rd /S /Q filelist
+if exist *.sh del *.sh
+if exist *.pl del *.pl
+if exist *.txt del *.txt
+popd
+pushd centos6\x64
+SET COMPONENT=CentOS 6 Build 64-bit
+echo [clean]: %COMPONENT%
+call vagrant destroy -f
+if exist build rd /S /Q build
+if exist filelist rd /S /Q filelist
+if exist *.sh del *.sh
+if exist *.pl del *.pl
+if exist *.txt del *.txt
 popd
 goto quit
 :error
